DRY up BlacklistFilter and WhitelistFilter by introducing AbstractPipelinesFilter
