+++ server/src/main/java/com/thoughtworks/go/server/domain/user/WhitelistFilter.java
@@ -17,33 +17,19 @@
 package com.thoughtworks.go.server.domain.user;
 
 import com.thoughtworks.go.config.CaseInsensitiveString;
-import com.thoughtworks.go.presentation.pipelinehistory.StageInstanceModel;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
-public class WhitelistFilter extends PipelinesFilter implements DashboardFilter {
-    private final String name;
-
+public class WhitelistFilter extends AbstractPipelinesFilter {
     public WhitelistFilter(String name, List<CaseInsensitiveString> pipelines, Set<String> state) {
-        super(state, pipelines);
-        this.name = name;
+        super(name, pipelines, state);
     }
 
     public List<CaseInsensitiveString> pipelines() {
         return Collections.unmodifiableList(pipelines);
-    }
-
-    @Override
-    public String name() {
-        return name;
-    }
-
-    @Override
-    public Set<String> state() {
-        return state;
     }
 
     @Override
@@ -62,12 +48,12 @@ public class WhitelistFilter extends PipelinesFilter implements DashboardFilter 
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         WhitelistFilter that = (WhitelistFilter) o;
-        return Objects.equals(name, that.name) &&
+        return Objects.equals(name(), that.name()) &&
                 Objects.equals(pipelines, that.pipelines);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name, pipelines);
+        return Objects.hash(name(), pipelines);
     }
 }
